--- backend/models/gameMode/blitz.js
@@ -1,9 +1,0 @@
-class Blitz  extends GameMode {
-    constructor() {
-        super()
-    }
-
-    getWinCondition(){
-        return true;
-    }
-}
