--- backend/models/gameMode/gameMode.js
@@ -1,10 +1,0 @@
-class GameMode {
-    constructor(){
-
-    }
-
-    getWinCondition(){
-        return true;
-    }
-}
-
