+++ backend/models/gameModes/Blitz.js
@@ -1,0 +1,13 @@
+const GameMode = require('./GameMode');
+
+class Blitz extends GameMode {
+    constructor() {
+        super()
+    }
+
+    getWinCondition(){
+        return true;
+    }
+}
+
+module.exports = Blitz;
