+++ backend/models/gameModes/Serie.js
@@ -1,0 +1,13 @@
+const GameMode = require('./GameMode')
+
+class Serie extends GameMode {
+    constructor(){
+        super()
+    }
+
+    getWinCondition(){
+        return true;
+    }
+}
+
+module.exports = Serie;
